+++ assembly/src/ast/instruction/opcode.rs
@@ -114,11 +114,17 @@ pub enum OpCode {
     U32Clo,
     U32Cto,
     U32Lt,
+    U32LtImm,
     U32Lte,
+    U32LteImm,
     U32Gt,
+    U32GtImm,
     U32Gte,
+    U32GteImm,
     U32Min,
+    U32MinImm,
     U32Max,
+    U32MaxImm,
 
     // ----- stack manipulation ------------------------------------------------------------------
     Drop,